requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
discord==1.7.3
discord-components==2.1.2
discord.py==1.7.3
dnspython==2.1.0
idna==3.3
importlib-metadata==4.9.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
motor==2.5.1
multidict==5.2.0
Pillow==8.4.0
pycparser==2.21
pymongo==3.12.1
PyNaCl==1.4.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0